x/crypto/acme: use a HEAD-safe URL for nonce fetch
This is a minimal fix to the HEAD rate limit problem discussed in
golang/go#18428. We send the HEADs for fetching nonces to a URL that we
know will be safer to hit than the POST-only resources where Let's
Encrypt is interpreting HEAD requests as POSTs and incrementing rate
limits as such. The safe URL chosen is the directory URL of the ACME
server.

A larger fix might use nonces gathered from previous API calls instead
of making a second API call for each API call made. This also does not
address the rate limit error being hidden by the nonce error as
discussed in golang/go#18428.

Updates golang/go#18428.

Change-Id: I378c64759bd5315aa96b4daff107d2741d742349